requirements.txt
pyjwt
cryptography
requests